fixing tasks 4-6

+++ lesson_1/task_5/main.cpp
@@ -5,9 +5,9 @@ using namespace std;
 
 int main()
 {
-    float mass, fuel;
+    int mass, fuel;
     cout << "enter ZAZ spaceship's mass, kg: ";
     cin >> mass;
-    fuel = round(mass/3.0)-2;
+    fuel = round(mass/3)-2;
     cout << "fuel needed: " << fuel << " kg.";
 }
